divided commands into subcommands with each file for main category
